requirements.txt
cvzone==1.5.2
mediapipe==0.8.7.3
bnumpy==1.19.5
opencv-python==4.5.3.56
